Refactor + more RDE Type behavior methods. Tests pass
